requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
pypdf
langchain_google_genai